Création composant ArticlePost Permet de poster un nouvel article
